+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 Also see [API_CHANGELOG](API_CHANGELOG.md).
 
+* Changed method of pruning segments from least-recently-used to a heuristic based on the synapse permanences.
+
 ## 2.1.0
 * REST API for htm.core
 * Ubuntu 20.04 LTS supported